+++ src/cpu/opcodes/stack.rs
@@ -26,7 +26,7 @@ impl super::super::Cpu {
 
         self.registers
             .f
-            .set_carry(super::is_overflow_8(self.stack_pointer as u16, e as u16));
+            .set_carry(super::is_overflow_8(self.stack_pointer, e as u16));
         self.registers.f.set_half_carry(super::is_half_carry_8(
             self.stack_pointer as u8,
             e as u8,
